+++ src/components/TopBar/TopBar.tsx
@@ -306,7 +306,8 @@ class TopBar extends React.Component <{
 								<Icon>help</Icon>
 							</IconButton>
 						</Tooltip>
-						<IconButton color='inherit' onClick={() => {this.setState({settingsDialog: true})}} component="span">
+						<IconButton color='inherit' onClick={() => {this.setState({settingsDialog: true})}} component="span"
+							style={{display: 'none'}}>
 							<Icon>settings</Icon>
 						</IconButton>
 					</Toolbar>