+++ CHANGELOG.md
@@ -6,9 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.7.0] - 02/09/2021
+
 ### Added
 
-- Settings dialog to change semester
+- Settings dialog to change semester (hidden for now)
 
 ### Fixed